Remove call to scan() in AccessPointList
Calling scan on wpa_cli with the interface in AP mode wipes the current
list of results. As this web interface is typically used in that mode,
we don't want to do that. We've removed the call to scan and assume that
the calling code will do that if the need arises.